Split up topics selectors into different files.
