+++ app/modules/topics/index.js
@@ -2,7 +2,7 @@
 
 import actions from './actions';
 import * as model from './model';
-import * as selectors from './selectors';
+import selectors from './selectors';
 import components from './components';
 
 const topics = {
