+++ app/modules/topics/saga/task/save.js
@@ -8,14 +8,14 @@ import contentItems from 'modules/contentItems';
 import actions from '../../actions';
 import * as a from '../../actionTypes';
 import * as m from '../../model';
-import { getById } from '../../selectors';
+import selectors from '../../selectors';
 
 const saveSaga = function* (action: a.SaveContentAction): Saga<void> {
   const {
     id,
   } = action.payload;
 
-  const topic: m.Topic = yield select(getById, { id });
+  const topic: m.Topic = yield select(selectors.getById, { id });
   const contentItemDescendants = yield select(
     contentItems.selectors.getSelfAndAllDescendantsById,
     { id: topic.rootContentItemId },
