+++ app/modules/topics/selectors/index.js
@@ -1,0 +1,15 @@
+// @flow
+
+import getAll from './getAll';
+import getAllById from './getAllById';
+import getAllTopicIdsByUserId from './getAllTopicIdsByUserId';
+import getById from './getById';
+
+const selectors = {
+  getAll,
+  getAllById,
+  getAllTopicIdsByUserId,
+  getById,
+};
+
+export default selectors;
